+++ test/acceptance/tests/basic/terraform/basic-install/main.tf
@@ -49,7 +49,7 @@ variable "launch_type" {
 variable "consul_ecs_image" {
   description = "Consul ECS image to use."
   type        = string
-  default     = "public.ecr.aws/hashicorp/consul-ecs:0.4.1"
+  default     = "docker.mirror.hashicorp.services/hashicorppreview/consul-ecs:0.4.1-dev"
 }
 
 variable "server_service_name" {
@@ -221,6 +221,8 @@ EOT
   outbound_only     = true
   // This keeps the application running for 10 seconds.
   application_shutdown_delay_seconds = 10
+  // Test with a port other than the default of 20000.
+  envoy_public_listener_port = 21000
 
   tls                            = var.secure
   consul_server_ca_cert_arn      = module.consul_server.ca_cert_arn